int -> NSInteger, timer states kept in an enum and moved with after event types to timer.h, cleanups
git-svn-id: svn://svn.freaque.net/enigma2@420 a295cab0-163a-0410-857a-8207b47ee9d6

+++ Classes/RCButton.h
@@ -11,9 +11,9 @@
 
 @interface RCButton : UIButton {
 @public
-	int rcCode;
+	NSInteger rcCode;
 }
 
-@property (nonatomic) int rcCode;
+@property (nonatomic) NSInteger rcCode;
 
 @end
